99.5 %

+++ app/src/main/java/com/example/doancn/Models/SubmissionX.kt
@@ -2,7 +2,7 @@ package com.example.doancn.Models
 
 data class SubmissionX(
     val dateCreated: String,
-    val fileId: Int,
+    val fileId: Long,
     val late: Boolean,
     val name: String,
     val student : User,
